requirements.txt
numpy>=1.24
pandas>=2.0
requests>=2.28
pytz>=2023.3
scipy>=1.10
python-dotenv>=1.0
pytest>=7.0
matplotlib>=3.7
pyarrow>=14.0
websocket-client>=1.6
